ovaltine: Include device-specific kernel fragment
For now the one and only purpose is to select the kernel config
that filters out DTBs that are not device-specific.

It surely helps to limit the number of DTBs that get built during
kernel compilation, but the main goal is to workaround for the
DTB[O]s merging issue that arises when adding all DTBs present
in vendor/oplus subdir. The issue could probably be solved with some
additional logic in Qualcomm's merge_dtbs script, but let's just
keep it untouched.

Change-Id: I72fae0fc99dc468e61eba3608d2326dcd118c0fd